chore(deps): update to latest version of KW SDK
Update to latest version of the kubewarden SDK, plus some other library
updates.

This is a major change, because the new SDK is no leveraging easyjson
and allows to interact with native k8s objects.

This commits performs some code changes, to ensure the new SDK can
be used. The code is still using the gjson library, because it allows
for a more compact code to be written.

Everything is green in terms of tests and e2e tests.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>